+++ services/json2capnp/transition_capnp_data/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "transition_capnp_data"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+json = "0.12"
+capnp = "0.14"
+serde_json = "1.0"
+uuid = "0.8"
+polyline = "0.9"
+geo = "0.17.0"
+geojson = "0.22.0"
+geobuf = "0.1"
+protobuf = "2.22.0"
+regex = "1.5.5"
