+++ services/json2capnp/transition_capnp_data/src/capnp/_generating_classes.txt
