+++ services/json2capnp/transition_capnp_data/src/capnp/householdCollection_capnp.rs
