+++ services/json2capnp/transition_capnp_data/src/capnp/household_capnp.rs
