+++ services/json2capnp/transition_capnp_data/src/enum_mappings.rs
