+++ services/json2capnp/transition_capnp_data/src/serialization/mod.rs
@@ -1,0 +1,21 @@
+/*
+ * Copyright 2025 Polytechnique Montreal and contributors
+ *
+ * This software may be used and distributed according to the terms of the
+ * GNU General Public License version 2 or any later version.
+ *
+ */
+
+/* Define and expose all data types in the module */
+
+pub mod agency_collection;
+pub mod data_source_collection;
+pub mod line_collection;
+pub mod line;
+pub mod node_collection;
+pub mod node;
+pub mod od_trip_collection;
+pub mod path_collection;
+pub mod place_collection;
+pub mod scenario_collection;
+pub mod service_collection;
